enigma2 (20120327 rel32 -> 20120706 rel32) ------------------------------------------ - updated translations - some storage manager optimization - fixed possible epgcache crash - fixed teletext subtitle crash - optimized teletext subtitle render quality - fixed common interface assignment plugin bug - fixed teletext subtitle selection screen - fixed skipping backward from live to timeshift - added expert option "Rotor is exclusively controlled   by this dreambox" to fix slow zapping times between   rotor and non-rotor lnbs (enabled by default) - fixed tuner resource leak on signal loss - fixed displaying of multilingual epg descriptions - improved transponder search speed (skip transponders without usable services) - fixed possible crash on enigma2 shutdown - fixed scrollabel for texts longer than a whole page - updated default favourite list and astra default services - fixed possible problem with multiple CIs for the same service/provider/caid

+++ usr/lib/enigma2/python/Plugins/SystemPlugins/CommonInterfaceAssignment/plugin.py
@@ -582,7 +582,7 @@ def activate_all(session):
 
 					for caid in slot.findall("caid"):
 						read_caid = caid.get("id").encode("UTF-8")
-						usingcaid.append(long(read_caid,16))
+						usingcaid.append(int(read_caid,16))
 
 					for service in slot.findall("service"):
 						read_service_ref = service.get("ref").encode("UTF-8")
